move wrapper types to enum files and delete old code
