+++ investment_length_unit_type.py
@@ -15,3 +15,18 @@ class InvestmentLengthUnitType(Enum):
 
     def __str__(self):
         return self._name_
+    
+class InvestmentLengthUnitWrapper:
+    investment_length_unit_type: InvestmentLengthUnitType
+
+    def __init__(self, input_string):
+        self.investment_length_unit_type = self._input_value_to_investment_length_unit_type(input_string)
+    
+    def _input_value_to_investment_length_unit_type(self, string):
+        if string == '1':
+            return InvestmentLengthUnitType.MONTHS
+        elif string == '2':
+            return InvestmentLengthUnitType.QUARTERS
+        elif string == '3':
+            return InvestmentLengthUnitType.YEARS
+        raise ValueError(f"Invalid input.  There is no length type associated with the input value: {string}")
